added basic support for midi control. the I2S bus outputs a frequency corresponding to the received midi key at the given velocity

+++ USB_DEVICE/App/usbd_midi_if.c
@@ -9,8 +9,8 @@
 #include "usbd_midi_if.h"
 #include "usbd_midi.h"
 
-__weak void MIDI_CC_Received(uint8_t channel, uint8_t code, uint16_t value) {}
-__weak void MIDI_Note_On_Received(uint8_t channel, uint8_t key, uint16_t velocity) {}
+__weak void MIDI_CC_Received(uint8_t channel, uint8_t code, uint8_t value) {}
+__weak void MIDI_Note_On_Received(uint8_t channel, uint8_t key, uint8_t velocity) {}
 __weak void MIDI_Note_Off_received(uint8_t channel, uint8_t key) {}
