+++ USB_DEVICE/App/usbd_midi_if.h
@@ -16,8 +16,8 @@
 
 extern USBD_MIDI_ItfTypeDef USBD_MIDI_Interface_fops_FS;
 
-void MIDI_CC_Received(uint8_t channel, uint8_t code, uint16_t value);
-void MIDI_Note_On_Received(uint8_t channel, uint8_t key, uint16_t velocity);
+void MIDI_CC_Received(uint8_t channel, uint8_t code, uint8_t value);
+void MIDI_Note_On_Received(uint8_t channel, uint8_t key, uint8_t velocity);
 void MIDI_Note_Off_received(uint8_t channel, uint8_t key);
